Extract register page into component
